+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/VisualMessageBoxForm.cs
@@ -187,7 +187,10 @@ internal partial class VisualMessageBoxForm : KryptonForm
                     SystemSounds.Exclamation.Play();
                     break;
                 case KryptonMessageBoxIcon.Shield:
-                    _messageIcon.Image = UACShieldIconResources.UAC_Shield_Windows_11;
+                    // Use theme-aware approach instead of hardcoded Windows 11 resource
+                    var currentPaletteMode = KryptonManager.CurrentGlobalPaletteMode;
+                    var shieldIcon = UacShieldIconHelper.GetThemeAwareShieldIcon(currentPaletteMode);
+                    _messageIcon.Image = shieldIcon?.ToBitmap() ?? UACShieldIconResources.UAC_Shield_Windows_10;
                     break;
                 case KryptonMessageBoxIcon.WindowsLogo:
                     // Because Windows 11 displays a generic application icon,
@@ -267,9 +270,10 @@ internal partial class VisualMessageBoxForm : KryptonForm
                     SystemSounds.Exclamation.Play();
                     break;
                 case KryptonMessageBoxIcon.Shield:
-                    _messageIcon.Image = OSUtilities.IsWindowsTen
-                        ? UACShieldIconResources.UAC_Shield_Windows_10
-                        : UACShieldIconResources.UAC_Shield_Windows_7;
+                    // Use theme-aware approach instead of hardcoded OS detection
+                    var currentPaletteMode = KryptonManager.CurrentGlobalPaletteMode;
+                    var shieldIcon = UacShieldIconHelper.GetThemeAwareShieldIcon(currentPaletteMode);
+                    _messageIcon.Image = shieldIcon?.ToBitmap() ?? UACShieldIconResources.UAC_Shield_Windows_10;
                     break;
                 case KryptonMessageBoxIcon.WindowsLogo:
                     _messageIcon.Image = OSUtilities.IsWindowsTen
