+++ Source/Krypton Components/Krypton.Toolkit/General/Definitions.cs
@@ -3139,32 +3139,4 @@ public enum KryptonEmojiListType
 
 #endregion
 
-#region Enum ShieldIconThemeMode
-
-/// <summary>
-/// Defines how UAC shield icons should handle theme changes.
-/// </summary>
-public enum ShieldIconThemeMode
-{
-    /// <summary>
-    /// Automatically detect and adapt to the current theme.
-    /// </summary>
-    Automatic = 0,
-
-    /// <summary>
-    /// Force light theme appearance regardless of system theme.
-    /// </summary>
-    Light = 1,
-
-    /// <summary>
-    /// Force dark theme appearance regardless of system theme.
-    /// </summary>
-    Dark = 2,
-
-    /// <summary>
-    /// Use the system's default theme behavior.
-    /// </summary>
-    System = 3
-}
-
-#endregion
+// ShieldIconThemeMode enum removed - now using theme-aware approach
