+++ Source/Krypton Components/Krypton.Toolkit/General/GlobalStaticValues.cs
@@ -104,11 +104,6 @@ public class GlobalStaticValues
     /// <summary>The public emoji list URL</summary>
     public const string DEFAULT_PUBLIC_EMOJI_LIST_URL = @"https://unicode.org/Public/draft/emoji/emoji-test.txt";
 
-    public const int SIID_SHIELD = 77;
-    public const int SHGSI_ICON = 0x000000100;
-    public const int SHGSI_LARGEICON = 0x000000000;
-    public const int SHGSI_SMALLICON = 0x000000001;
-
     #region Arrays
 
     #region Images
